requirements.txt
numpy
matplotlib
scikit-image
pillow